DOC: Added `Getting started' section.
- Also made zeropad return the same type of array.

+++ amitgroup/util/misc.py
@@ -8,6 +8,6 @@ def zeropad(data, padwidth):
         padwidth = (padwidth,)*len(shape) 
         
     padded_shape = map(lambda ix: ix[1]+padwidth[ix[0]]*2, enumerate(shape))
-    new_data = np.zeros(padded_shape)
+    new_data = np.zeros(padded_shape, dtype=data.dtype)
     new_data[ [slice(w, -w) if w > 0 else slice(None) for w in padwidth] ] = data 
     return new_data
